+++ Citadel/src/utopia/citadel/database/factory/organization/DeletionCancelFactory.scala
@@ -5,7 +5,7 @@ import utopia.flow.datastructure.immutable.{Constant, Model}
 import utopia.flow.generic.ValueUnwraps._
 import utopia.metropolis.model.partial.organization.DeletionCancelData
 import utopia.metropolis.model.stored.organization.DeletionCancel
-import utopia.vault.nosql.factory.FromValidatedRowModelFactory
+import utopia.vault.nosql.factory.{FromRowFactoryWithTimestamps, FromValidatedRowModelFactory}
 
 /**
   * Used for reading organization deletion cancellations from DB
@@ -13,11 +13,14 @@ import utopia.vault.nosql.factory.FromValidatedRowModelFactory
   * @since 16.5.2020, v1.0
   */
 object DeletionCancelFactory extends FromValidatedRowModelFactory[DeletionCancel]
+	with FromRowFactoryWithTimestamps[DeletionCancel]
 {
 	// IMPLEMENTED	-------------------------------
 	
 	override def table = Tables.organizationDeletionCancellation
 	
+	override def creationTimePropertyName = "created"
+	
 	override protected def fromValidatedModel(model: Model[Constant]) = DeletionCancel(model("id"),
-		DeletionCancelData(model("deletionId"), model("creatorId")))
+		DeletionCancelData(model("deletionId"), model("creatorId"), model(creationTimePropertyName)))
 }