+++ Metropolis/src/utopia/metropolis/model/partial/organization/DeletionCancelData.scala
@@ -3,6 +3,9 @@ package utopia.metropolis.model.partial.organization
 import utopia.flow.datastructure.immutable.Model
 import utopia.flow.generic.ModelConvertible
 import utopia.flow.generic.ValueConversions._
+import utopia.flow.time.Now
+
+import java.time.Instant
 
 /**
   * Contains basic information about an organization deletion cancellation
@@ -11,7 +14,9 @@ import utopia.flow.generic.ValueConversions._
   * @param deletionId Id of the cancelled deletion
   * @param creatorId Id of the user who cancelled the deletion (if known)
   */
-case class DeletionCancelData(deletionId: Int, creatorId: Option[Int] = None) extends ModelConvertible
+case class DeletionCancelData(deletionId: Int, creatorId: Option[Int] = None, created: Instant = Now)
+	extends ModelConvertible
 {
-	override def toModel = Model(Vector("deletion_id" -> deletionId, "creator_id" -> creatorId))
+	override def toModel = Model(Vector(
+		"deletion_id" -> deletionId, "creator_id" -> creatorId, "created" -> created))
 }