+++ Vault/src/utopia/vault/nosql/factory/LinkedFactory.scala
@@ -2,23 +2,18 @@ package utopia.vault.nosql.factory
 
 import utopia.flow.datastructure.immutable.{Constant, Model}
 import utopia.vault.model.immutable.Row
-import utopia.vault.sql.JoinType
 
 import scala.util.{Failure, Try}
 
 /**
- * Used for converting DB data to models for tables that contain a single link to another table / model
- * @author Mikko
- * @since 21.8.2019, v1.3.1+
+ * Used for converting DB data to models for tables that contain a single link to another table / model.
+  * Expects there to always exist a link between the two tables
+ * @author Mikko Hilpinen
+ * @since 21.8.2019, v1.3.1
  */
-trait LinkedFactory[+Parent, Child] extends FromRowFactory[Parent]
+trait LinkedFactory[+Parent, Child] extends FromRowFactory[Parent] with LinkedFactoryLike[Parent, Child]
 {
 	// ABSTRACT	---------------------
-	
-	/**
-	 * @return Factory used for parsing linked child data
-	 */
-	def childFactory: FromRowFactory[Child]
 	
 	/**
 	 * Parses model & child data into a complete parent model. Parsing failures are handled by
@@ -32,7 +27,7 @@ trait LinkedFactory[+Parent, Child] extends FromRowFactory[Parent]
 	
 	// IMPLEMENTED	-----------------
 	
-	override def joinType = JoinType.Inner
+	override def isAlwaysLinked = true
 	
 	override def apply(row: Row) =
 	{
@@ -41,6 +36,4 @@ trait LinkedFactory[+Parent, Child] extends FromRowFactory[Parent]
 		else
 			Failure(new NoModelDataInRowException(childFactory.table, row))
 	}
-	
-	override def joinedTables = childFactory.tables
 }