+++ Vault/src/utopia/vault/nosql/factory/PossiblyMultiLinkedFactory.scala
@@ -1,7 +1,5 @@
 package utopia.vault.nosql.factory
 
-import utopia.vault.sql.JoinType
-
 /**
   * Factory which retrieves items that are linked to 0-n other items of a certain type
   * @author Mikko Hilpinen
@@ -9,5 +7,5 @@ import utopia.vault.sql.JoinType
   */
 trait PossiblyMultiLinkedFactory[+Parent, Child] extends MultiLinkedFactory[Parent, Child]
 {
-	override def joinType = JoinType.Left
+	override def isAlwaysLinked = false
 }