+++ Metropolis/Changes.md
@@ -1,4 +1,9 @@
 # Utopia Metropolis - List of Changes
 
+## v1.0.1 (in development)
+### Other Changes
+- Added a few models from the **Exodus Module**
+- **DeletionCancel** now contains property `.created`
+
 ## v1.0 - 17.4.2021
 Initial release. See README.md for a list of main features.